pyproject.toml
[project]
name = "bex"
dynamic = ["version"]
description = "bex"
readme = "README.md"
authors = [
    { name = "Lucino772"}
]
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "ruamel-yaml>=0.18.10,<1",
    "stdlibx-cancel==0.1.0",
    "stdlibx-compose==0.1.0",
    "stdlibx-option==0.2.0",
    "stdlibx-result==0.2.0",
    "typer>=0.16.0,<1",
]

[dependency-groups]
dev = [
    "ruff>=0.15.1",
    "pyinstaller>=6.17.0",
]

[project.scripts]
bex = "bex:main"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/bex/_version.py"
template = """
# file generated by setuptools_scm
# don't change, don't track in version control
__all__ = ["__version__", "__version_tuple__"]
__version__: str = {version!r}
__version_tuple__: tuple[int | str, ...] = {version_tuple!r}
"""

[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 80

[tool.ruff.lint.isort]
known-first-party = ["bex"]

[tool.ruff.lint]
extend-select = ["I", "RUF022", "RUF023", "W505", "TC", "FA", "C4", "PYI", "COM818"]
extend-ignore = ["D203", "D213"]

[tool.ruff.lint.pycodestyle]
max-doc-length = 80

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"
